Create new golden_trays app and refactor model, views, and templates into the new app directory
