+++ golden_trays/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GoldenTraysConfig(AppConfig):
+    name = 'golden_trays'
