+++ golden_trays/fixtures/initial_crop_attributes.json
