+++ golden_trays/templates/add_crop_attributes.html
