+++ inventory/admin.py
@@ -1,12 +1,7 @@
 from django.contrib import admin
-from inventory.models import Crop, CropAttribute, CropAttributeOption, CropRecord, Slot, Variety, SanitationRecord, WeekdayRequirement, InventoryAction, KillReason, CropGroup
+from inventory.models import Variety, SanitationRecord, WeekdayRequirement, InventoryAction, KillReason, CropGroup
 from orders.models import TrayType, MicrogreenSize, Order, Product, LiveCropProduct, HarvestedCropProduct
 
-admin.site.register(Crop)
-admin.site.register(CropAttribute)
-admin.site.register(CropAttributeOption)
-admin.site.register(CropRecord)
-admin.site.register(Slot)
 admin.site.register(Variety)
 admin.site.register(SanitationRecord)
 admin.site.register(WeekdayRequirement)
